fixes for code review & OpenMP/Serial bug fix of non-polymorphic call was used
