+++ src/occa/internal/bin/occa.cpp
@@ -3,24 +3,13 @@
 #include <stdexcept>
 
 #include <occa/core.hpp>
-
 #include <occa/internal/bin/occa.hpp>
 #include <occa/internal/utils/env.hpp>
-#include <occa/internal/lang/modes/serial.hpp>
-#include <occa/internal/lang/modes/openmp.hpp>
-#include <occa/internal/lang/modes/cuda.hpp>
-#include <occa/internal/lang/modes/hip.hpp>
-#include <occa/internal/lang/modes/opencl.hpp>
-#include <occa/internal/lang/modes/metal.hpp>
-#include <occa/internal/lang/modes/dpcpp.hpp>
+#include <occa/internal/io.hpp>
+#include <occa/internal/utils/sys.hpp>
 #include <occa/internal/modes.hpp>
 
-#ifdef BUILD_WITH_OCCA_TRANSPILER
 #include <occa/internal/utils/transpiler_utils.h>
-#include "oklt/pipeline/normalizer_and_transpiler.h"
-#include "oklt/util/io_helper.h"
-#include "oklt/core/error.h"
-#endif
 
 #include <memory>
 
@@ -135,208 +124,6 @@ namespace occa {
       return true;
     }
 
-#ifdef BUILD_WITH_OCCA_TRANSPILER
-    namespace v3 {
-
-        bool runTranslate(const json &options,
-                          const json &arguments,
-                          const json &kernelProps,
-                          const std::string &originalMode,
-                          const std::string &mode)
-        {
-            static const std::map<std::string, oklt::TargetBackend> targetBackends =
-            {
-                {"openmp", oklt::TargetBackend::OPENMP},
-                {"cuda", oklt::TargetBackend::CUDA},
-                {"hip", oklt::TargetBackend::HIP},
-                {"dpcpp", oklt::TargetBackend::DPCPP},
-            };
-
-            const bool printLauncher = options["launcher"];
-            const std::string filename = arguments[0];
-
-            if (!io::exists(filename)) {
-                printError("File [" + filename + "] doesn't exist" );
-                ::exit(1);
-            }
-
-            auto transpiler = targetBackends.find(mode);
-            if(transpiler == targetBackends.end()) {
-                printError("Unable to translate for mode [" + originalMode + "]");
-                ::exit(1);
-            }
-
-            auto defines = transpiler::buildDefines(kernelProps);
-            auto includes = transpiler::buildIncludes(kernelProps);
-            auto hash = transpiler::getKernelHash(kernelProps);
-
-            std::filesystem::path sourcePath = io::expandFilename(filename);
-            auto sourceCode = oklt::util::readFileAsStr(sourcePath);
-            if(!sourceCode) {
-                printError("Can't open file: " + sourcePath.string());
-                ::exit(sourceCode.error());
-            }
-            oklt::UserInput input {
-                .backend = transpiler->second,
-                .source = std::move(sourceCode.value()),
-                .headers = {},
-                .sourcePath = sourcePath,
-                .includeDirectories = std::move(includes),
-                .defines = std::move(defines),
-                .hash = std::move(hash)
-            };
-            auto result = normalizeAndTranspile(std::move(input));
-
-            if(!result) {
-                std::stringstream ss;
-                for(const auto &err: result.error()) {
-                    ss << err.desc << std::endl;
-                }
-                printError(ss.str());
-                ::exit(1);
-            }
-
-            if (options["verbose"]) {
-                json translationInfo;
-                // Filename
-                translationInfo["translate_info/filename"] = io::expandFilename(filename);
-                // Date information
-                translationInfo["translate_info/date"] = sys::date();
-                translationInfo["translate_info/human_date"] = sys::humanDate();
-                // Version information
-                translationInfo["translate_info/occa_version"] = OCCA_VERSION_STR;
-                translationInfo["translate_info/okl_version"] = OKL_VERSION_STR;
-                // Kernel properties
-                translationInfo["kernel_properties"] = kernelProps;
-
-                io::stdout
-                    << "/* Translation Info:\n"
-                    << translationInfo
-                    << "*/\n";
-            }
-
-            auto userOutput = result.value();
-            bool hasLauncher = transpiler->second == oklt::TargetBackend::CUDA ||
-                               transpiler->second == oklt::TargetBackend::HIP ||
-                               transpiler->second == oklt::TargetBackend::DPCPP;
-            if(printLauncher && hasLauncher) {
-                io::stdout << userOutput.launcher.source;
-            } else {
-                io::stdout << userOutput.kernel.source;
-            }
-
-            return true;
-        }
-    }
-#endif
-
-    namespace v2 {
-        bool runTranslate(const json &options,
-                          const json &arguments,
-                          const json &kernelProps,
-                          const std::string &originalMode,
-                          const std::string &mode)
-        {
-            using ParserBuildFunc = std::function<std::unique_ptr<lang::parser_t>(const json &params)>;
-            static const std::map<std::string, ParserBuildFunc> originalParserBackends =
-                {
-                    {"", [](const json &params) {
-                         return std::make_unique<lang::okl::serialParser>(params);
-                     }},
-                    {"serial", [](const json &params) {
-                         return std::make_unique<lang::okl::serialParser>(params);
-                     }},
-                    {"openmp", [](const json &params) {
-                         return std::make_unique<lang::okl::openmpParser>(params);
-                     }},
-                    {"cuda", [](const json &params) {
-                         return std::make_unique<lang::okl::cudaParser>(params);
-                     }},
-                    {"hip", [](const json &params) {
-                         return std::make_unique<lang::okl::hipParser>(params);
-                     }},
-                    {"opencl", [](const json &params) {
-                         return std::make_unique<lang::okl::openclParser>(params);
-                     }},
-                    {"metal", [](const json &params) {
-                         return std::make_unique<lang::okl::metalParser>(params);
-                     }},
-                    {"dpcpp", [](const json &params) {
-                         return std::make_unique<lang::okl::dpcppParser>(params);
-                     }}
-                };
-
-            const bool printLauncher = options["launcher"];
-            const std::string filename = arguments[0];
-
-            if (!io::exists(filename)) {
-                printError("File [" + filename + "] doesn't exist" );
-                ::exit(1);
-            }
-
-            auto parserIt = originalParserBackends.find(mode);
-            if(parserIt == originalParserBackends.end()) {
-                printError("Unable to translate for mode [" + originalMode + "]");
-                ::exit(1);
-            }
-
-            std::unique_ptr<lang::parser_t> parser = parserIt->second(kernelProps);
-            parser->parseFile(filename);
-
-            bool success = parser->succeeded();
-            if (!success) {
-                ::exit(1);
-            }
-
-            if (options["verbose"]) {
-                json translationInfo;
-                // Filename
-                translationInfo["translate_info/filename"] = io::expandFilename(filename);
-                // Date information
-                translationInfo["translate_info/date"] = sys::date();
-                translationInfo["translate_info/human_date"] = sys::humanDate();
-                // Version information
-                translationInfo["translate_info/occa_version"] = OCCA_VERSION_STR;
-                translationInfo["translate_info/okl_version"] = OKL_VERSION_STR;
-                // Kernel properties
-                translationInfo["kernel_properties"] = kernelProps;
-
-                io::stdout
-                    << "/* Translation Info:\n"
-                    << translationInfo
-                    << "*/\n";
-            }
-
-            if (printLauncher && ((mode == "cuda")
-                                  || (mode == "hip")
-                                  || (mode == "opencl")
-                                  || (mode == "dpcpp")
-                                  || (mode == "metal"))) {
-                lang::parser_t *launcherParser = &(((occa::lang::okl::withLauncher*) parser.get())->launcherParser);
-                io::stdout << launcherParser->toString();
-            } else {
-                io::stdout << parser->toString();
-            }
-            return true;
-        }
-    }
-
-    int getTranspilerVersion(const json &options) {
-        json jsonTranspileVersion = options["transpiler-version"];
-        int transpilerVersion = 2;
-        //INFO: have no idea why json here has array type
-        if(jsonTranspileVersion.isArray()) {
-            json elem = jsonTranspileVersion.asArray()[0];
-            if(elem.isString()) {
-                try {
-                    transpilerVersion = std::stoi(elem.string());
-                } catch(const std::exception &)
-                {}
-            }
-        }
-        return transpilerVersion;
-    }
-
     bool runTranslate(const json &args) {
       const json &options = args["options"];
       const json &arguments = args["arguments"];
@@ -349,15 +136,7 @@ namespace occa {
       kernelProps["defines"].asObject() += getOptionDefines(options["define"]);
       kernelProps["okl/include_paths"] = options["include-path"];
 
-#ifdef BUILD_WITH_OCCA_TRANSPILER
-      int transpilerVersion = getTranspilerVersion(options);
-      if(transpilerVersion > 2) {
-        return v3::runTranslate(options, arguments, kernelProps, originalMode, mode);
-      }
-      return v2::runTranslate(options, arguments, kernelProps, originalMode, mode);
-#else
-      return v2::runTranslate(options, arguments, kernelProps, originalMode, mode);
-#endif
+      return transpiler::runTranspiler(options, arguments, kernelProps, originalMode, mode);
     }
 
     bool runCompile(const json &args) {
@@ -378,7 +157,7 @@ namespace occa {
       kernelProps["verbose"] = kernelProps.get("verbose", true);
       kernelProps["okl/include_paths"] = options["include-path"];
       kernelProps["defines"].asObject() += getOptionDefines(options["define"]);
-      kernelProps["transpiler-version"] = getTranspilerVersion(options);
+      kernelProps["transpiler-version"] = transpiler::getTranspilerVersion(options);
 
       device device(deviceProps);
       device.buildKernel(filename, kernelName, kernelProps);
