+++ src/occa/internal/core/launchedDevice.hpp
@@ -70,7 +70,7 @@ namespace occa {
     ) = 0;
     //==================================
 
-#ifdef BUILD_WITH_OCCA_TRANSPILER
+#ifdef BUILD_WITH_CLANG_BASED_TRANSPILER
     private:
         bool transpileFile(const std::string &filename,
                            const std::string &outputFile,
