+++ src/occa/internal/modes/openmp/device.cpp
@@ -6,11 +6,11 @@
 #include <occa/internal/modes/openmp/device.hpp>
 #include <occa/internal/modes/openmp/utils.hpp>
 
-#ifdef BUILD_WITH_OCCA_TRANSPILER
+#ifdef BUILD_WITH_CLANG_BASED_TRANSPILER
 #include <occa/internal/utils/transpiler_utils.h>
-#include "oklt/pipeline/normalizer_and_transpiler.h"
-#include "oklt/core/error.h"
-#include "oklt/util/io_helper.h"
+#include <oklt/pipeline/normalizer_and_transpiler.h>
+#include <oklt/core/error.h>
+#include <oklt/util/io_helper.h>
 #endif
 
 namespace occa {
@@ -32,7 +32,7 @@ namespace occa {
       );
     }
 
-#ifdef BUILD_WITH_OCCA_TRANSPILER
+#ifdef BUILD_WITH_CLANG_BASED_TRANSPILER
     bool device::transpileFile(const std::string &filename,
                        const std::string &outputFile,
                        const occa::json &kernelProps,
