+++ src/occa/internal/modes/openmp/device.hpp
@@ -12,11 +12,11 @@ namespace occa {
       std::string lastCompiler;
       std::string lastCompilerOpenMPFlag;
 
-#ifdef BUILD_WITH_OCCA_TRANSPILER
+#ifdef BUILD_WITH_CLANG_BASED_TRANSPILER
       bool transpileFile(const std::string &filename,
                          const std::string &outputFile,
                          const occa::json &kernelProps,
-                         lang::sourceMetadata_t &metadata);
+                         lang::sourceMetadata_t &metadata) override;
 #endif
 
     public:
