+++ src/occa/internal/utils/transpiler_utils.h
@@ -7,11 +7,20 @@
 namespace occa {
 namespace transpiler {
 
+int getTranspilerVersion(const json &options);
+
 std::string getKernelHash(const json &kernelProp);
 std::vector<std::string> buildDefines(const json &kernelProp);
 std::vector<std::filesystem::path> buildIncludes(const json &kernelProp);
 void makeMetadata(lang::sourceMetadata_t &sourceMetadata,
                   const std::string &jsonStr);
+
+bool runTranspiler(const json &options,
+                   const json &arguments,
+                   const json &kernelProps,
+                   const std::string &originalMode,
+                   const std::string &mode);
+
 
 }
 }
